project/requirements.txt
alembic==1.11.1
asyncpg==0.28.0
fastapi==0.100.0
sqlmodel==0.0.8
uvicorn==0.22.0
boto3
opencv-python
numpy
python-multipart
python-dotenv